requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
openai==1.3.5
pypdf==3.17.1
python-docx==0.8.11
email-validator==2.1.0
Flask==3.1.3
Flask_Bcrypt==1.0.1
flask_cors==6.0.2
passlib==1.7.4
python_jose==3.5.0